fix: remove not necessary prop and move layout
